app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pytest==8.3.3
python-jose==3.3.0
pydantic[email]==2.9.2
passlib==1.7.4
pydantic-settings==2.5.2
psycopg2-binary==2.9.9
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.2
langgraph==0.2.35
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
ipython==8.28.0
bs4==0.0.2
